fix(examples):  fix #warning causing examples build to fail.
When the docs build, the "Build examples" step is done by compiling
the code examples with "everything turned on".  That means that

are both set to 1.

Simultaneously, in `lv_font.h`, there is a

for both of those fonts, but Visual Studio WILL NOT build LVGL because
MSVC does not yet support the #warning preprocessor directive.  MSVC
considers it an error and so does not even generate the output object files.

However

still works, and is a legitimate way to get the message across for
developers using the CJK fonts to switch over to the new font without
breaking their builds.

This handles both problems with 2 lines of code.