docs(table css):  fix table formatting...
This `td` element styling was causing table cells to all be center-aligned.
When we tried to introduce a table with a lot of cell content, it was
both unsightly and not very readable.

As a previous professional typesetter, I can say (with some authority)
that I have never seen the need for default table-cell content to be
center aligned.  Default browser/CSS table-alignment behavior confirms
this.

Further, I have checked all the existing LVGL-documentation tables
(listed below) and removing this behavior does not adversely affect
any of them -- they still look good (if not enhanced).

If center-aligned text for tables are needed in the future they can
be achieved by giving the specific table a `:class:` parameter and
specificially styling that one table in our `custom.css` file.

The existing LVGL-documentation tables are:

- `./docs/intro/introduction.rst`:303
- `./docs/details/integration/chip/renesas/supported_boards.rst`:5
- `./docs/details/integration/os/yocto/lvgl_recipe.rst`:289
- `./docs/details/integration/os/rt-thread.rst`:54
- `./docs/details/common-widget-features/coordinates.rst`:65