docs:  (custom.css):  add caption feature and...
make immediate use of it in `fs.rst`.  This gives a "hint" for the
reader on how the image may be managed to see it better and
zoom in on its parts and move around after being zoomed in.

Also added a long-overdue extra space above headings of
various sizes.  <h1> is not in that list because <h1> is always
at the top of the page, and so should not have extra space
above it.