docs(custom.css):  partially address code-highlighting problem...
...reading code blocks in dark mode.  Specifically, the original problem
we wanted to address was that function names and argument names
in the example code blocks had identical colors.  Unfortunately, the
thing parsing it classifies the 2 exactly the same, calling them "names"
(which is class="n" in Pygments).  However, it was observed that if
at least the punctuation was a different color, it would somewhat
improve readability.  And this was doable because the parser classifies
these "spans" as class="p" inside of a parent "highlight" class element.

In this PR, I took the apparent "LVGL-theme purple" color and lightened
it until it looked good in the other (fairly bright white) code elements.
The below illustrates the results.