docs(api_doc_builder.py):  fix TOC ordering on API pages.
Apparently the Python implementation of the `listdir()` function
differs between Windows and Linux.  I suspect it uses a system setting
under Windows where under "normal" circumstances, it provides a
sorted list, whereas under Linux, it comes out in directory-entry order
(typically creation sequence).  So sorting is introduced before the
lists (especially .h-file list) so that when the TOCs are generated, they
are in alphabetical order.

Fixes #8383