chore(lv_indev_t::feedback_cb):  remove outdated material...
As correctly pointed out in Issue #8514, the Feedback section of
`indev.rst` no longer applies because the `feedback_cb` field no longer
exists since v9.0.0.

Further, I found a comment in `lv_obj_event.c` related to this same
topic that no longer applies to the code it "labels".

Original code (v8.4):

```c
    /*Call the input device's feedback callback if set*/
    if(indev_act) {
        if(indev_act->driver->feedback_cb) indev_act->driver->feedback_cb(indev_act->driver, e->code);
        if(e->stop_processing) return LV_RES_OK;
        if(e->deleted) return LV_RES_INV;
    }
```

v9.0.0 code:

```c
    /*Call the input device's feedback callback if set*/
    if(indev_act) {
        if(e->stop_processing) return LV_RES_OK;
        if(e->deleted) return LV_RES_INV;
    }
```

As you can see, the code was removed, but the comment remained, probably
because it was placed over the whole IF block instead of over the line that
it referred to.

I removed the comment.

Fixes: #8514