fix(preprocess_lv_conf_internal.py):  replace `which` with...
...a cross-platform compatible solution to find `pcpp` executable.
When `-DLV_BUILD_SET_CONFIG_OPTS=1` was specified on the
command line, the script was using `which` to get the path to
`pcpp` executable, which on windows, outputs something like
`/c/Python/Python312/Scripts/pcpp` which is not a valid
Windows path.  It works under `bash`, but not from a Python
script.  The changes herein resolve that in a cross-platform
manner.

co-author:  André