docs(list formatting):  fix erroneous list formatting
A recent investigation into Definition-List formatting (should be
encapsulated in HTML `<dl>` elements) discovered that in a number of
places in the docs that LISTS and SUB-LISTS (should be in `<ul>`
elements) were being incorrectly formatted as Definition Lists rather
than Unordered Lists.

In most cases, this was caused by having a sub-list start without a
preceeding blank line, and in some cases also involved incorrect
indentation of the sub-list, which caused the sub-list to be interpreted
as a "literal block", which is an error.

This PR fixes all of the issues found with this problem.

The resulting correctly-formatted docs can be found
[here](http://crystal-clear-research.com/lvgl-docs-demo/master/#).

The IN-correctly-formatted docs can be found
[here](https://docs.lvgl.io/master/index.html).