docs(custom.css):  fix text color on view-on-github example buttons
A recent increase in the specificity of styling rules for LINKs
(`<a>` elements) caused this rule to be overridden, losing the color
and the other styling it provided for the [View on GitHub] buttons that
come with each example.  The result is that the text color is the LINK
color (purple), whereas with this increase in specificity, it correctly
overrides that color applying the styling styling originally intended
for that button.

It should probably be investigated what else got unintentionally
overridden by the `#furo-main-content` additions.