docs(custom.css):  fix arrows in code blocks
Previously, wherever the combination of "->" or "<-" appeared
using the monospaced font, the font itself was translating the
'>' or '<' into an arrow instead of an angle bracket.

This modification remedies that.  I do not believe that ligatures
(combinations of characters) are used anywhere else, since the
docs are in English.  Translations that use ligatures may need to
apply this attribute less globally.