fix(lv_windows_display):  fix window-size bug
Various versions of the Windows `CreateWindow()` functions do not, in all
cases, use the `width` and `height` arguments to govern the "drawing area"
(called "Client Area" in Windows vocabulary).  The `AdjustWindowRect()`
and `AdjustWindowRectEx()` functions (used when there is ALSO an
"extended window style" involved, which there is) are made to make
window size adjustments so that the Client Area (drawing area) DOES
turn out to have the desired dimensions.  This function is not currently
being used prior to calling the `CreateWindow()` function, so when
"Simulator Mode" is not used with the `lv_windows_create_display()`
function, the client area comes out considerably smaller than it should.

This PR implements `AdjustWindowRectEx()` as it was intended to be
used to adjust the `width` and `height` parameters of the
`CreateWindow()` function so that the drawing area has the desired
dimensions.

Fixes:  #9033