docs(custom.js):  fix enumerator expansion upon API page links
Issue #9043 correctly observed that enumerator values, when clicked in
a doc, wound the reader up at the bottom of the page, with the target
enumerator value hidden.  The reason for this was that the parent `enum`
<dl class="cpp enum unexpanded"> element was unexpanded, and thus
its child <dd> element (containing the target enumerator documentation)
remained hidden on the page.

This PR, examines the parent <dl> element, and if it is indeed unexpanded,
then it is expanded, allowing the browser to move to and show the
expanded enumerator as originally intended.

Fixes #9043