docs(substitutions.txt):  factor out repeated substitution defs
Throughout the docs are special symbols that are implemented through
a substitution definition that typically looks like this:

.. |deg|    unicode:: U+000B0 .. DEGREE SIGN

for the degree symbol.  There are about a dozen others.

This PR factors all of these out into a central location where they are
simply defined ONCE and then included where needed with an
`.. include:: /include/substitutions.txt` directive.