docs(macro api links):  fix links to macros in API pages
In some cases (e.g. when a list of API-page macros was long), clicking
on a :c:macro:\`MACRO_NAME\` in a document resulted in the browser
jumping to the bottom of the API page instead of showing the expanded
macro name in the page.  Adding `cppListing.scrollIntoView()` to the
code block dealing with that expanded documentation remedies this.