fix(spinbox):  fix stack corruption
This PR fixes:

- Fixes a design error that was causing stack corruption.  See below.
- Fixes 7 compiler warnings that were being generated under MSVC.
- Makes the 3rd argument and description of `lv_spinbox_set_digit_format()`
  consistent with the `dec_point_pos` argument of
  `lv_spinbox_set_dec_point_pos()` with the same meaning, to be clear in
  the API that they indeed have the same meaning.
- Makes `dec_point_pos` range-checking logic consistent between
  `lv_spinbox_set_dec_point_pos()` and `lv_spinbox_set_digit_format()`
  functions.
- Makes range-checking of the 2 range values consistent between:
  - lv_spinbox_set_range()
  - lv_spinbox_set_min_value()
  - lv_spinbox_set_max_value()
- Make `LV_SPINBOX_MAX_DIGIT_COUNT` have real meaning.  It is currently
  10 as set in `lv_spinbox.h`, but this is an automatic digit-count limit
  of the `int32_t` which is the type of the Spinbox's `value` as well as
  both range values.  Assuming this value should have meaning in the
  code, range-checking logic in the range-setting functions was modified
  to use it to generate range-clamping code if it is ever < 10.

### Cause of Stack Corruption

Calling `lv_spinbox_set_range()` with range values that required more
than the current `digit_count`, caused `lv_spinbox_updatevalue()` calls
(made after every setting change) to violate array boundaries by using a
negative index on the `digits[]` character array, overwriting whatever
was just before it on the stack.

Lower-level cause:  `lv_spinbox_updatevalue()` assumes `digit_count`
and the 2 range values are in a consistent state.  When they are not in
a consistent state, array boundaries get violated.  This is an acceptable
design assumption as long as no public API functions can cause them to
be left in an inconsistent state.

### Design Error:

Lower-level cause:  the following functions were allowing `digit_count`
and the 2 range values to be left in an inconsistent state.

- lv_spinbox_set_range()
- lv_spinbox_set_min_value()
- lv_spinbox_set_max_value()
- lv_spinbox_set_digit_count()
- lv_spinbox_set_digit_format()

Solution:

Fix those functions to never leave `digit_count` and the 2 range values
in an inconsistent state:

- Extend `digit_count` when needed when setting range values.
- Clamp range values when needed when setting `digit_count`.

Fixes #9255