docs(API pages):  convert to workable source link
Fixes #9393

Previously the GitHub source link icon led to a link that looked
like this:
```
https://github.com/lvgl/lvgl/blob/master/docs/src/API/core/lv_obj_h.rst?plain=true
```
which is an `.rst` file which ONLY exists in the intermediate directory
while docs are being built, which led to a 404 error since the file  doesn't
exist in the set of web-hosting files.  Now the link looks like this:
```
https://github.com/lvgl/lvgl/blob/master/src/core/lv_obj.h
```
which leads to the `.h` file from which it was derived.

This template does what the original template did, except that it
examines the resulting URL and adjusts it when it is from an API page.